Add new DataGridManager class
Introduces a new class to manage multiple DataGrids. The manager covers
an area set with world_size and splits this area up to by covered by a
number of DataGrids specified with datagrid_count. The cell_size
parameter is used (together with the world_size and the datagrid_count
parameter) to calculate the resulting size_in_cells of the DataGrid(s).
For now it only holds a single DataGrid and datagrid_count cant be
changed. In the future this DataGrid will be replaced by a Dictionary
or Array to hold multiple DataGrids and the class will be extended by
methods to manage them.