Bugfix
The method add_grid_at_pos translates the coordinates of the cells of
the other_grid to the local coordinates. After determining the
intersection, the local value has to be added to the according value
of the other_grid. This can be done by substracting the offset of the
other_grid from the local coordinates.
The offset of the intersection has been substrated so far which caused
unexpected behavior when adding an other grid near the left and top
side. This has been fixed.