Add new DataGridTemplate Class
When entering data into DataGrids radiate_value_at_position can be used
to generate values for cells based on their distance to the center cell
and a curve. Getting the distance involves calculating a square root for
each cell. To reduce the calculation time templates of commonly used
patterns can be prepared and stored. The method add_grid_at_pos can then
be used to add the data of a template to another grid, involving only
additions.
This commit introduces the DataGridTemplate, a wrapper for DataGrid. Also
adds two new methods to DataGridManager "initialize_templates" and
"get_template". To create a range of templates and to retrieve them.