Add DataGridComponent Class
Introduces a class called DataGridComponent that inherits from the
Node2D class. It has a radius and an Array of layers it is supposed to
be entered in. Layers are used to split data in different categories
like teams (enemy / friend). Every DataGridComponent joins the group
"DataComponents" when instatiated.
This class can be added as a child to any Node2D following it
automatically. The DataGridManager can just walk the
DataComponents-Group and register / de-register the Components to a
DataGrid by adding a template of a matching radius at the position of
the component.
Parameters used on registration are stored for deregistration.