Update DataGridManager to use new classes
DataGridManager now stores a DataGridCollection instead of a single
DataGrid covering the complete area. This make it possible to create
DataGridManager covering big areas without storing to many empty
DataGrids in unpopulated areas. There can be multiple DataGrids at
each position, allowing to keep unrelated data in seperate DataGrids.
Methods refering to the single DataGrid have been updated as well.
Additionally the DataGridManager now has an update method. It updates
the DataGrids by iterating through all the nodes in the
"DataComponents" group, registering and de-registering them to their
specified layers. At the end the update method emits a signal to notify
the visualization GridSprite of the changes.
DataGridManager also received methods to convert a world_position (in
pixels) to a DataGrid position and cell coordinates in the DataGrid-
Collection.