Move DataGridManager visualization to GridSprite
DataGridManager emits a signal when changing its size. GridSprite reacts
to this signal by updating the width of the grid it is showing. The
width is needed to get the values of the flat Array back to a 2D
representation (x + y * width).
DataGridManager should now set as a child of a GridSprite node. This
allows moving the reaction to the updated signal from the Main script to
the GridSprite script.