Add DataGridHub Singleton
Introduces a Singleton called DataGridHub to the project. This class
has an Array with references to all DataGridComponents. It allows
DataGridComponents to register themselves on instantiation.
DataGridManagers can loop over this Array to update their DataGrids with
data from the DataGridComponents.
When using a Group registered data of DataGridComponents that have been
freed, can't be deregistered. Unlike a Group, the Singleton can be used
to store RefCounted Objects. Those can be used to deregister data of a
freed DataGridComponent.