Add DataGridCompRef
Previously, the DataGridHub stored all DataGridComponents in an Array.
The DataGridManager traversed this Array to register and deregister the
Components in the DataGrids. However, once a registered Component was
freed, it was immediately removed from the DataGridHub's Array, making
it impossible to deregister their last values.
This commit introduces a new class named DataGridCompRef - the name
needs revision in the future. DataGridCompRef inherits from RefCounted,
ensuring that DataGridCompRefs are only freed when the last reference to
them is unreferenced.
DataGridComponents now have a DataGridCompRef to store all the data
previously held directly by the DataGridComponent. Instead of passing a
reference to themselves, they provide a reference to their CompRef to
the DataGridHub. Consequently, the DataGridHub maintains an Array of
DataGridCompRefs.
The DataGridManagers iterate through this Array, updating their
respective DataGrids with the data stored in the CompRefs. For each
CompRef, they begin by deregistering any previously registered data.
Before registering the current status, they check if the component that
holds the DataGridCompRef is still valid. If it is not, implying that
the component was freed at some point in the past, the DataGridCompRef
becomes unreferenced and subsequently freed as well. However, if the
DataGridComponent remains valid, its current status is registered.