Optimize radiate_value method
The radiate_value_at_position method of the DataGrid class should set
the values of grid cells inside a given radius. To check if cells are
within the given radius a distance test is performed. To avoid checking
the distance to every cells of the DataGrid, a bounding box is defined
first. Only the distance of cells inside the square bounding box are
compared to the radius.
To further improve the test, the square root needed to get the distance
can be avoided when checking if a if cell is inside the circle. The
condition `distance < radius` can be changed to `distance² < radius²`
because both numbers are always positive.
Cells towards the corners of the bounding box will fail the distance
check. The MathCurve shouldn't calculate a value in this case (as it did
until now), it only applies to cells inside the circle. Those corner
cells will be ignored.
The distance will only be calculated for cells that passed the
`is_inside_circle` test. MathCurve deals with values ranging from
0.0 - 1.0. So, the distance relative to the radius is passed to
calculate the value.
Lastly to get the index of the cell inside the `Array data` the formula
`x + y * size_in_cells.x` is used. The `y * size_in_cells.x` is moved
to the outer loop to avoid re-calculating it for every column.