New method add_from_pos_in_grid
Introduces a new method that grabs a region from another grid and adds
it into this grid. The regions has the same size as the grid calling the
method and is centered at the specified position.
The existing method `add_grid_centered_at_pos` can be used to add a
small grid at a certain position into a large grid. Like registering an
explosion somewhere in the world grid.
This new method can be used to extract a small region from a large grid.
This is usefull when an agent wants to look at its surrounding area
while not taking the whole world grid into consideration.