Update demo
Visualization of grids is now done via a custom `_draw()` function. This
may be less performant than using a shader but it is more flexible. The
size of the DataGrid is not limited to 4096 anymore and can be changed
at runtime.
This commit also introduces a prototype for the IntelligenceController
on the Entity class. The IntelligenceController creates a DataGrid
representing the area of the world that the Entity is interested in. It
can than extract values from the DataGridManager like the proximity of
other entities (in this case the FollowMouse Object). This DataGrid can
be evaluated by finding the cell with the highest or lowest value.
The DataGrid can be modified further by using interest_templates. In
this case a template is multiplied with the DataGrid to favor cells that
are closer to the Entity.