Support multiple DataGrids on DataGridManager
This commit extends he functionality of DataGridManager to support
multiple DataGrids laid out in a grid. This allows for more efficient
memory utilization, particularly in large maps, by enabling the addition
of DataGrids on demand and the removal of unused ones.
Key Changes:
- Introduce the `datagrid_count` parameter to specify the number of
columns and rows in the grid of DataGrids.
- Implement new methods for de/-registering and extracting data from an
area the of the DataGridManager. These handle scenarios where the area
of interest overlaps with multiple DataGrids within the DataGridManager.
The `get_touched_datagrids` method returns index offsets of the
Datagrids overlapping with the area of interest. The
`add_datagrid_centered_to_collection` method handles adding a single
DataGrid into multiple DataGrids on the DataGridManager. The
`add_into_datagrid_from_collection` method handles the aggregation of
data from multiple DataGrids into a single one.
- The DataGrid methods `add_grid_centered_at_pos`and
`add_from_pos_in_grid` have been changed to accept a new offset
parameter, which adjusts the `p_position` parameter to align with
neighboring maps.