Add registerable to DataGridCompfRef
Adds the bool registerable on DataGridCompfRef. It indicates whether a
DataGridCompfRef needs to be de/registered on the manager.
Additionally, adds a timer to DataGridComponent with the floats
time_since_registration and registration_frequency. When enough time
has passed since the last registration, the component checks if a
meaningful change hast occured that is worth de/registering. In that
case registerable on DataGridCompfRef is set to true.