Autonomous updating
Moved functionality of the update method to the _process method.
Previously, an update would be executed in reaction to the timeout
signal of a Timer. Instead, the DataGridManager walks the Array of
DataGridCompfRefs to find one that `is_registerable` and update it. The
number of DataGridCompfRefs that can be updated each frame can be set.
This allows to split the work load between multiple frames.
Some deprecated parameters and methods were removed.
The ordering of the methods has been changed to allign with the other
files.