Rename DataGridTemplate to InfluenceMapTemplate
Also makes InfluenceMapTemplate a private struct in DataGriddManager
instead of a class (inheriting from Resource) registered to Godot. This
fits the InfluenceMapTemplate better, since no other class except the
manager is using Templates.
InfluenceMapTemplate are stored in a LocalVector instead of a
TypedArray and the DataGridManager methods `initialize_templates` and
`get_template` have been adjusted to use the updated data structures.