Add mark_cells_outside_boundaries method
Introduces `mark_cells_outside_boundaries` method to DataGridManager
class. This method uses a new InfluenceMapTemplate with all values set
to 1.0.
Entities that intend to remain within the boundaries managed by the
DataGridManager can utilize this method by adding the template with a
negative magnitude to regions within their interest_map.
This is necessary because when combining influence from the Manager in
an interest_mapall areas of the interest_map that that do not intersect
with the Manager's covered area will remain at 0.0. When normalizing the
interest_map, those zeros can become desirable targets.

The `mark_cells_outside_boundaries` method can be used with a
magnitude of -2 after an Entity has combined and normalized all layers
of influence into an interest_map and normalized it.