Add influence_type member variable
Introduces influence_type member variable which is an integer. Can be
used in combination with an enum to represent different kinds of
influence behave in the same way. The proximity of an entity is uses
a rather small radius and decreases linearly while the area of interest
is usually bigger and its fall off could be calculated by a polynomial
function.