Store different types of templates in HashMap
The create_template method of DataGridManager is updated to take a
influence_type and a MathCurve in addition to the radius parameters.
This allows to create a number of InfluenceMapTemplates with the given
curve and store them inside a LocalVector. The LocalVector is then added
to a HashMap with its influence_type as the key. This can be used with a
enum to name the different types depending on the needs of the game.
The get_template is updated to retrieve a template from the new data
structure.