Set InterestMap sized based on MoveSpeed
Export an entity variable on IntelligenceController to be able to get
the move_speed to calculate the size of the InterestMap. This way a
slowed entity will only consider closer targets while a entity with a
move_speed bonus will look for targets that are further away.
Also, entities should not avoid their own proximity (causing endless
random running) when avoiding the proximity of entities on their team.
Instead of using `radiate_value_at_position` to zero out their own
proximity in their interest map, a template is added.